// psa.service.sampletrackingservice/src/models/bloodSample.ts
/*
 * SPDX-FileCopyrightText: 2021 Helmholtz-Zentrum für Infektionsforschung GmbH (HZI) <PiaPost@helmholtz-hzi.de>
 *
 * SPDX-License-Identifier: AGPL-3.0-or-later
 */

export interface BloodSample {
  id: number;
  user_id: string;
  sample_id: string | null;
  blood_sample_carried_out: boolean | null;
  remark: string | null;
}